+++ Backend/controllers/programController.js
@@ -29,8 +29,8 @@ const getProgramById = (req, res) => {
 
 //get specific user programs -- viewed in Profile 
 const getProgramHistory = (req, res) => {
-    const username = req.params.username;
-    programModel.getProgramHistory(username, (err, result) => {
+    const userId = req.params.userId;
+    programModel.getProgramHistory(userId, (err, result) => {
         if (!err) {
             return res.status(200).json(result);
         } else {
@@ -74,15 +74,31 @@ const updateProgram = (req, res) => {
     if (req.file) {
         program.image = req.file.path; // Store the path of the uploaded image
     }
-    programModel.updateProgram(program, (err, result) => {
-        if (!err) {
-            if(result.affectedRows == 0){
-                return res.status(404).json({message: "Program id not found"});
-            }
-            return res.status(200).json({ message: "Program updated successfully" });
-        } else {
-            return res.status(500).json(err); 
+
+    programModel.getProgramById(program.id, (err, result) => {
+        if (err) {
+            return res.status(500).json(err);
         }
+
+        if (result.length == 0) {
+            return res.status(404).json({ message: "Program id not found" });
+        }
+
+        // Ownership check
+        if (result[0].userId !== res.userLocal.userId) {
+            console.log("this program", result[0].userId, "logged in as", res.userLocal.userId)
+            return res.status(403).json({ message: "Forbidden" });
+        }
+        programModel.updateProgram(program, (err, result) => {
+            if (!err) {
+                if(result.affectedRows == 0){
+                    return res.status(404).json({message: "Program id not found"});
+                }
+                return res.status(200).json({ message: "Program updated successfully" });
+            } else {
+                return res.status(500).json(err); 
+            }
+        });
     });
 };
 
@@ -90,25 +106,41 @@ const updateProgram = (req, res) => {
 const deleteProgram = (req, res) => {
     const id = req.params.id;
     const imagePath = req.params.imagePath; 
-
-    console.log(imagePath);
-    programModel.deleteProgram(id, (err, result) => {
-        if (!err) {
-            if (result.affectedRows === 0) {
-                return res.status(404).json({ message: "Program id not found" });
-            }
-
-            // Delete the corresponding image file from the 'uploads/' directory
-            fs.unlink(`uploads/${imagePath}`, (err) => {
-                if (err) {
-                    console.error(err);
-                    return res.status(500).json({ message: "Error deleting image file" });
-                }
-                return res.status(200).json({ message: "Program and corresponding image deleted successfully" });
-            });
-        } else {
+    
+    programModel.getProgramById(id, (err, result) => {
+        if (err) {
             return res.status(500).json(err);
         }
+
+        if (result.length == 0) {
+            return res.status(404).json({ message: "Program id not found" });
+        }
+
+        // Ownership check
+        if (result[0].userId !== res.userLocal.userId) {
+            console.log("this program", result[0].userId, "logged in as", res.userLocal.userId)
+            return res.status(403).json({ message: "Forbidden" });
+        }
+
+        // console.log(imagePath);
+        programModel.deleteProgram(id, (err, result) => {
+            if (!err) {
+                if (result.affectedRows === 0) {
+                    return res.status(404).json({ message: "Program id not found" });
+                }
+
+                // Delete the corresponding image file from the 'uploads/' directory
+                fs.unlink(`uploads/${imagePath}`, (err) => {
+                    if (err) {
+                        console.error(err);
+                        return res.status(500).json({ message: "Error deleting image file" });
+                    }
+                    return res.status(200).json({ message: "Program and corresponding image deleted successfully" });
+                });
+            } else {
+                return res.status(500).json(err);
+            }
+        });
     });
 };
 
@@ -127,7 +159,7 @@ const registerProgram = (req, res) => {
 //get program registrant
 const getRegistrant = (req, res) => {
     const programId = req.params.programId;
-    console.log('Received programId:', programId);
+    // console.log('Received programId:', programId);
     programModel.getRegistrant(programId, (err, result) => {
         if (err) {
             console.error('Database query error:', err);
